% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11eigentest-package.R
\docType{package}
\name{cpp11blazetest-package}
\alias{cpp11blazetest}
\alias{cpp11blazetest-package}
\title{cpp11blazetest: Testing cpp11blaze}
\description{
Testing cpp11blaze. I run this within a tiny package to test the compiled code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/cpp11blaze}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11blaze/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
